+++ HandyNotes_WarWithin/handler/handler.lua
@@ -523,7 +523,7 @@ end
 ns.render_string = render_string
 ns.render_string_list = render_string_list
 
-local npc_texture, follower_texture, currency_texture, junk_texture
+local npc_texture, follower_texture, currency_texture, junk_texture, notable_npc_texture
 local icon_cache = {}
 local trimmed_icon = function(texture)
     if not icon_cache[texture] then
@@ -537,18 +537,25 @@ local trimmed_icon = function(texture)
     end
     return icon_cache[texture]
 end
-local atlas_texture = function(atlas, extra, crop)
+local atlas_texture = function(atlas, extra, left, right, top, bottom)
     atlas = C_Texture.GetAtlasInfo(atlas)
     if type(extra) == "number" then
         extra = {scale=extra}
     end
-    if crop then
-        local xcrop = (atlas.rightTexCoord - atlas.leftTexCoord) * crop
-        local ycrop = (atlas.bottomTexCoord - atlas.topTexCoord) * crop
-        atlas.rightTexCoord = atlas.rightTexCoord - xcrop
-        atlas.leftTexCoord = atlas.leftTexCoord + xcrop
-        atlas.bottomTexCoord = atlas.bottomTexCoord - ycrop
-        atlas.topTexCoord = atlas.topTexCoord + xcrop
+    if left and not right then
+        -- this is the "trim every side by this" path
+        right = 1 - left
+        top = left
+        bottom = 1 - left
+    end
+    if left then
+        -- An atlas is already cropped into a texture, so we need to treat something else as our 1
+        local horizontal = atlas.rightTexCoord - atlas.leftTexCoord
+        local vertical = atlas.bottomTexCoord - atlas.topTexCoord
+        atlas.rightTexCoord = atlas.leftTexCoord + (right * horizontal)
+        atlas.leftTexCoord = atlas.leftTexCoord + (left * horizontal)
+        atlas.bottomTexCoord = atlas.topTexCoord + (bottom * vertical)
+        atlas.topTexCoord = atlas.topTexCoord + (top * vertical)
     end
     return ns.merge({
         icon = atlas.file,
@@ -557,6 +564,11 @@ local atlas_texture = function(atlas, extra, crop)
 end
 ns.atlas_texture = atlas_texture
 local default_textures = {
+    --[[
+    note to self:
+    atlas_texture("delves-scenario-treasure-unavailable", nil, 0, 0.9, 0.1, 1)
+    atlas_texture("delves-scenario-treasure-available", nil, 0, 0.9, 0.05, 0.95)
+    --]]
     VignetteLoot = atlas_texture("VignetteLoot", 1.1),
     VignetteLootElite = atlas_texture("VignetteLootElite", 1.2),
     Garr_TreasureIcon = atlas_texture("Garr_TreasureIcon", 2.2),
@@ -675,9 +687,14 @@ local function work_out_texture(point)
     end
     if point.npc then
         if not npc_texture then
+            notable_npc_texture = atlas_texture("nazjatar-nagaevent", 1, 0.2)
             npc_texture = atlas_texture("DungeonSkull", 1)
         end
-        return npc_texture
+        if ns.db.show_npcs_emphasizeNotable and ns.PointIsNotable(point, true) then
+            return notable_npc_texture
+        else
+            return npc_texture
+        end
     end
     if point.currency then
         if not currency_texture then
@@ -1452,6 +1469,12 @@ function HL:OnInitialize()
     self.db = LibStub("AceDB-3.0"):New(myname.."DB", ns.defaults)
     ns.db = self.db.profile
     ns.hidden = self.db.char.hidden
+    -- Quick upgrade-cycle:
+    if ns.db.show_npcs_onlynotable then
+        ns.db.show_npcs_filter = "notable"
+        ns.db.show_npcs_onlynotable = nil
+    end
+
     -- Initialize our database with HandyNotes
     HandyNotes:RegisterPluginDB(myname:gsub("HandyNotes_", ""), HLHandler, ns.options)